+++ plugins/tracker-resources/src/components/issues/TitlePresenter.svelte
@@ -30,6 +30,7 @@
 <style lang="scss">
   .root {
     display: flex;
+    flex-grow: 1;
     min-width: 0;
     white-space: nowrap;
     overflow: hidden;