+++ plugins/tracker-resources/src/index.ts
@@ -57,6 +57,7 @@ import Views from './components/views/Views.svelte'
 import KanbanView from './components/issues/KanbanView.svelte'
 import tracker from './plugin'
 import { getIssueId, getIssueTitle } from './utils'
+import CreateIssue from './components/CreateIssue.svelte'
 
 export async function queryIssue<D extends Issue> (
   _class: Ref<Class<D>>,
@@ -144,12 +145,13 @@ export default async (): Promise<Resources> => ({
     KanbanView,
     TeamProjects,
     Roadmap,
-    IssuePreview
+    IssuePreview,
+    CreateIssue
   },
   completion: {
     IssueQuery: async (client: Client, query: string) => await queryIssue(tracker.class.Issue, client, query)
   },
   function: {
-    getIssueTitle
+    IssueTitleProvider: getIssueTitle
   }
 })