+++ changelog.md
@@ -2,21 +2,32 @@
 
 ## 0.6.29 (upcoming)
 
+Platform:
+
+- Object selector actions
+
 Tracker:
 
-- Remember view options 
-- Names of parent issues
+- Remember view options
 - My issues
+- Names of parent issues
 - Roadmap
+- Context menus (Priority/Status/Assignee)
 - 
 
 Chunter:
 
 - Reactions on messages
+- Priority filter
+- Context menu selector for state/assignee
 
 HR:
 
 - Leaves schedule
+
+Core:
+
+- Invite link expire from 1 hour
 
 ## 0.6.28
 
@@ -29,7 +40,6 @@ Tracker:
 - Issue state history.
 - Subissue issue popup.
 - Label support
-- Priority filter
 
 Lead: